#893 Add system property to disable closing broken connections

+++ src/main/org/firebirdsql/ds/FBXADataSource.java
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 package org.firebirdsql.ds;
 
+import org.firebirdsql.gds.JaybirdSystemProperties;
 import org.firebirdsql.gds.impl.GDSFactory;
 import org.firebirdsql.gds.impl.GDSType;
 import org.firebirdsql.jaybird.xca.*;
@@ -18,6 +19,7 @@ import javax.sql.XAConnection;
 import javax.sql.XADataSource;
 import java.io.Serial;
 import java.io.Serializable;
+import java.lang.System.Logger.Level;
 import java.sql.SQLException;
 
 /**
@@ -95,7 +97,12 @@ public class FBXADataSource extends FBAbstractCommonDataSource implements XAData
 
         @Override
         public void connectionErrorOccurred(XcaConnectionEvent ce) {
-            destroyConnection(ce);
+            boolean forceCloseOnFatal = JaybirdSystemProperties.isXaConnectionForceCloseOnFatal();
+            LOG.log(Level.TRACE, () -> "ConnectionErrorOccurred (forceCloseOnFatal=%b)".formatted(forceCloseOnFatal),
+                    ce.getException());
+            if (forceCloseOnFatal) {
+                destroyConnection(ce);
+            }
         }
 
         private void destroyConnection(XcaConnectionEvent ce) {
@@ -103,7 +110,7 @@ public class FBXADataSource extends FBAbstractCommonDataSource implements XAData
             try {
                 mc.destroy(ce);
             } catch (SQLException e) {
-                LOG.log(System.Logger.Level.WARNING, "Ignored exception closing unmanaged connection", e);
+                LOG.log(Level.WARNING, "Ignored exception closing unmanaged connection", e);
             } finally {
                 mc.removeConnectionEventListener(this);
             }
